Fix fseeko() call. Change prototype for ckmalloc().
git-svn-id: http://www.telegraphics.com.au/svn/psdparse/trunk@413 d127708d-35f1-0310-882e-b6c776866be2

+++ psdparse.h
@@ -299,7 +299,7 @@ void fatal(char *s);
 void warn_msg(char *fmt, ...);
 void alwayswarn(char *fmt, ...);
 #define checkmalloc(N) ckmalloc(N, __FILE__, __LINE__)
-void *ckmalloc(long n, char *file, int line);
+void *ckmalloc(size_t n, char *file, int line);
 void fputcxml(char c, FILE *f);
 void fputsxml(char *str, FILE *f);
 void fwritexml(char *buf, size_t count, FILE *f);
